chore: Update android (WIP)
Android build was wildly out of date. This is the culmination of a few
days of battling gradle version issues, kotlin version issues, and java
version issues. In this commit, viro react compiles with each of the
dependencies (renderer, etc.) and creates the bridge properly both in a
React Native project and an Expo project.

There's a "hack" in the Expo withViroAndroid.ts file. Expo generates a
MainApplication.kt file that doesn't actually add the ViroReact packages
to the application.

Todo:
- Continue fixing arrow functions in some class components
- Rebuild virocore and debug crash in VROMathInvertMatrix
- Deploy release
- Add code to show a view instead of initializing AR if the device
  doesn't support AR Core