chore: Update android (#261)
Android build was wildly out of date. This is the culmination of a few
days of battling gradle version issues, kotlin version issues, and java
version issues. In this commit, viro react compiles with each of the
dependencies (renderer, etc.) and creates the bridge properly both in a
React Native project and an Expo project.

There's a "hack" in the Expo withViroAndroid.ts file. Expo generates a
MainApplication.kt file that doesn't actually add the ViroReact packages
to the application.

This also fixes a crash caused by a message "cannot read property VIRO_VERSION of undefined."

require('expo') in a react native project caused this issue, but for
some reason the error message was about VIRO_VERSION, which was working
correctly.